+++ src/main/resources/api_help/add.help
@@ -1,0 +1,17 @@
+add <input>
+    %s
+
+    <input> is a json object representing the password to add to the passwords list.
+    It can contain four keys and their values: pass, site, user, other.
+
+    Examples:
+
+    ...add {pass=test123,site=gmail.com,user=John}
+    -> The input mustn't contain all four keys
+
+    Error codes:
+        20  The input wasn't adequate to the json format
+
+    Hint:
+        Error messages will be displayed as following:
+            !<code> <message>
