+++ src/test/java/PasswordTest.java
@@ -17,7 +17,7 @@ public class PasswordTest {
 
         File temp = File.createTempFile("passvault","bin");
 
-        String pass = PasswordGenerator.generatePassword(new PasswordGenerator.PassChars[]{PasswordGenerator.PassChars.BIG_LETTERS}, (int) (Math.random() * 30));
+        String pass = PasswordGenerator.generatePassword(new PasswordGenerator.PassChars[]{PasswordGenerator.PassChars.BIG_LETTERS}, (int) (Math.random() * 30) + 1);
 
         Password.savePasswords(passwords, temp,pass);
         assertEquals(passwords, Password.readPasswords(temp, pass));
